backend/requirements.txt
GitPython==3.1.40
langchain==0.0.335
langchainplus-sdk==0.0.20
openai==1.2.4
tiktoken==0.4.0
#azure-core==1.27.0
uvicorn==0.21.1
astor==0.8.1
mlfoundry==0.10.2
servicefoundry==0.9.24
fastapi==0.104.1
boto3==1.26.76
pdf2image==1.16.3
grpcio==1.59.2
grpcio-tools==1.59.2
qdrant-client==1.6.4
python-dotenv==1.0.0
spacy==3.5.3
transformers==4.30.1
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.5.0/en_core_web_md-3.5.0-py3-none-any.whl
pandas==2.1.3
PyYAML==6.0
torchvision==0.15.2
torch==2.0.1
timm==0.9.2
python-doctr==0.7.0
pdfplumber==0.9.0
deepdoctection==0.27
packaging==21.3
nltk==3.8.1
weaviate-client==3.25.3
pydantic==1.10.12
orjson==3.9.2